+++ packages/plugin-qiankun/src/common.test.ts
@@ -3,10 +3,9 @@
  * @since 2019-10-22
  */
 
-import { addSpecifyPrefixedRoute, testPathWithPrefix } from '../common';
-import { expectRoutes, originRoutes } from './mock';
+import { testPathWithPrefix } from './common';
 
-it('testPathPrefix', () => {
+test('testPathPrefix', () => {
   // browser history
   expect(testPathWithPrefix('/js', '/')).toBeFalsy();
 
@@ -106,17 +105,4 @@ it('testPathPrefix', () => {
   expect(testPathWithPrefix('#/js/:abc/jsx/:cde', '#/js/123/jsk')).toBeFalsy();
   expect(testPathWithPrefix('#/js/:abc?', '#/js')).toBeTruthy();
   expect(testPathWithPrefix('#/js/*', '#/js/245')).toBeTruthy();
-});
-
-it('testRecursiveCoverRouter', () => {
-  // 在原route的基础上添加指定路由单测
-  expect(String(addSpecifyPrefixedRoute(originRoutes, 'test'))).toEqual(
-    String(expectRoutes),
-  );
-  expect(String(addSpecifyPrefixedRoute(originRoutes, 'test', 'app2'))).toEqual(
-    String(expectRoutes),
-  );
-  expect(String(addSpecifyPrefixedRoute(originRoutes, true, 'test'))).toEqual(
-    String(expectRoutes),
-  );
 });
