+++ packages/plugin-qiankun/src/common.ts
@@ -3,21 +3,9 @@
  * @since 2019-06-20
  */
 
-import { cloneDeep } from 'lodash';
 import pathToRegexp from 'path-to-regexp';
-import { IRoute } from 'umi';
-import { SlaveOptions } from './types';
 
 export const defaultMountContainerId = 'root-subapp';
-
-// 主应用跟子应用的默认 root id 区分开，避免冲突
-export const defaultMasterRootId = 'root-master';
-export const defaultSlaveRootId = 'root-slave';
-
-export const defaultHistoryType = 'browser';
-
-export const qiankunStateForSlaveModelNamespace = '@@qiankunStateForSlave';
-export const qiankunStateFromMasterModelNamespace = '@@qiankunStateFromMaster';
 
 // @formatter:off
 export const noop = () => {};
@@ -49,39 +37,3 @@ export function testPathWithPrefix(pathPrefix: string, realPath: string) {
     testPathWithDynamicRoute(pathPrefix, realPath)
   );
 }
-
-const recursiveCoverRouter = (source: Array<IRoute>, nameSpacePath: string) =>
-  source.map((router: IRoute) => {
-    if (router.routes) {
-      recursiveCoverRouter(router.routes, nameSpacePath);
-    }
-    if (router.path !== '/' && router.path) {
-      return {
-        ...router,
-        path: `${nameSpacePath}${router.path}`,
-      };
-    }
-    return router;
-  });
-
-export const addSpecifyPrefixedRoute = (
-  originRoute: Array<IRoute>,
-  keepOriginalRoutes: SlaveOptions['keepOriginalRoutes'],
-  pkgName?: string,
-) => {
-  const copyBase = originRoute.filter(_ => _.path === '/');
-  if (!copyBase[0]) {
-    return originRoute;
-  }
-
-  const nameSpaceRouter: any = cloneDeep(copyBase[0]);
-  const nameSpace = keepOriginalRoutes === true ? pkgName : keepOriginalRoutes;
-
-  nameSpaceRouter.path = `/${nameSpace}`;
-  nameSpaceRouter.routes = recursiveCoverRouter(
-    nameSpaceRouter.routes,
-    `/${nameSpace}`,
-  );
-
-  return [nameSpaceRouter, ...originRoute];
-};
