+++ packages/plugin-qiankun/src/constants.ts
@@ -1,0 +1,5 @@
+export const defaultMasterRootId = 'root-master';
+export const defaultSlaveRootId = 'root-slave';
+export const defaultHistoryType = 'browser';
+export const qiankunStateForSlaveModelNamespace = '@@qiankunStateForSlave';
+export const qiankunStateFromMasterModelNamespace = '@@qiankunStateFromMaster';
