+++ packages/plugin-qiankun/src/slave/qiankunModel.ts.tpl
