+++ packages/plugin-qiankun/src/slave/slaveRuntimePlugin.ts.tpl
